remove operator
