+++ bin/node/cli/src/chain_spec.rs
@@ -7,7 +7,7 @@ use plasm_runtime::Block;
 use plasm_runtime::{
     BabeConfig, BalancesConfig, ContractsConfig, EVMConfig, EthereumConfig, GenesisConfig,
     GrandpaConfig, ImOnlineConfig, IndicesConfig, SessionConfig, SessionKeys, StakerStatus,
-    StakingConfig, SudoConfig, SystemConfig, VestingConfig, WASM_BINARY,
+    StakingConfig, SudoConfig, SystemConfig, VestingConfig, WASM_BINARY, PlasmRewardsConfig, PlasmValidatorConfig
 };
 use sc_chain_spec::ChainSpecExtension;
 use sc_service::ChainType;
@@ -204,6 +204,12 @@ fn make_genesis(
             .collect(),
         }),
         pallet_ethereum: Some(EthereumConfig {}),
+        pallet_plasm_rewards: Some(PlasmRewardsConfig {
+            ..Default::default()
+        }),
+        pallet_plasm_validator: Some(PlasmValidatorConfig {
+            validators: vec![]
+        }),
     }
 }
