+++ bin/node/runtime/Cargo.toml
@@ -32,7 +32,6 @@ sp-session = { version = "3.0.0", default-features = false }
 sp-staking = { version = "3.0.0", default-features = false }
 
 # frame dependencies
-pallet-contract-operator = { path = "../../../frame/operator", default-features = false }
 #pallet-operator-trading = { path = "../../../frame/trading", default-features = false }
 pallet-dapps-staking = { path = "../../../frame/dapps-staking", default-features = false }
 pallet-plasm-rewards = { path = "../../../frame/plasm-rewards", default-features = false }
